+++ test/setup-jsdom.js
@@ -1,16 +1,27 @@
 'use strict';
 
+import { URL, URLSearchParams } from 'url';
 import jsdom from 'jsdom';
 
 const dom = new jsdom.JSDOM();
 global.window = dom.window;
 global.document = dom.window.document;
+global.navigator = dom.window.navigator;
+global.NodeList = dom.window.NodeList;
 global.Element = dom.window.Element;
 global.Text = dom.window.Text;
-global.HTMLTextAreaElement = dom.window.HTMLTextAreaElement;
 global.HTMLDivElement = dom.window.HTMLDivElement;
+global.HTMLTextAreaElement = dom.window.HTMLTextAreaElement;
+global.HTMLSpanElement = dom.window.HTMLSpanElement;
+
 global.requestAnimationFrame = fn => setTimeout(fn, 16);
 global.cancelAnimationFrame = id => clearTimeout(id);
+
+global.URL = URL;
+global.URLSearchParams = URLSearchParams;
+global.location = new URL('https://www.bitbucket.org');
+
+global.Headers = function() {};
 
 global.$ = global.jQuery = require('jquery');